requirements.txt
fastapi
uvicorn
pandas
faiss-cpu
groq
numpy
streamlit
requests